requirements.txt
langchain
langchain_text_splitters
pymupdf
tiktoken
supabase
openai
tqdm
python-dotenv
regex